Update wording for "provides basic input" technical skill
I did feeback interviews with some folks who interpreted this skill
as "provides basic mentoring to their pair, and input on their pair's
career progress". I've updated the description to make it clearer that
this is a basic, early P1 level of contributing at a minimal level to
story progress.