Update the "leaving space for others" arc
* Move "actively leaves space for others to share their ideas to P2"
because P1s are usually still developing enough ideas to be comfortable
sharing them at all. P2 is still the "do it yourself" level of
contribution, so I think this skill makes sense there.
* Add a P3 skill "Actively works to ensure quieter teammates have room
to share and get recognition for their ideas". P3 is doing things that
have a whole team impact. One can start in this arc by thinking about
the amount of space you take up in the room (P2), and then start
noticing and influencing the amount of space others take up in the room.
* Add a P4 skill "Coaches pivots on how to facilitate conversations and
meetings that allow other quieter teammates to share and get recognition
for their ideas". Noticing how much space one takes up in the room, and
how much space others take up, and balancing that, isn't an easy skill
to develop. Coaching other pivots on how to do this skill would have a
big impact on the team, by enabling other pivots.

Relates to #issue-32