Make item explicitly about solo/small projects
It seems like this advanced decision-making skill is often applicable on
solo or very small projects. The wording implies that the person is making those
decisions by themselves, leading to some confusion. The previous skill
item seems more general, but it seems worth specifying for this one.
Solo exploration or prototypes are common advanced-level projects, so
this feels useful.